Hangman Game
This code was very fun to create. I learned a lot doing it and it was very rewarding when it finally worked. It started off as nearly 300 lines of code with errors in every turn. Eventually the code size was reduced to just over 100 lines and it seems to work just fine! I am sure it can be improved upon, if you have any suggestions please feel free to share!
Gonza